+++ config.go
@@ -30,6 +30,8 @@ type Config struct {
 	NetInterface   string
 	Layout         string
 	MaxLogSize     int64
+	ExportPort     string
+	Extensions     []string
 }
 
 func Parse(in io.Reader, conf *Config) error {
@@ -109,6 +111,10 @@ func Parse(in io.Reader, conf *Config) error {
 				return err
 			}
 			conf.MaxLogSize = int64(iv)
+		case "export":
+			conf.ExportPort = kv[1]
+		case "extensions":
+			conf.Extensions = strings.Split(kv[1], ",")
 		}
 	}