+++ CHANGELOG.md
@@ -23,17 +23,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   a histogram.
 - Temp widget displays degree symbol (merged from BartWillems, thanks
   also fleaz)
+- Support for (device) plugins, and abstracting devices from widgets. This
+  allows adding functionality without adding bulk.
 
 ### Fixed
 
 - Keys not controlling process widget, #59
+- The one-column bug, #62
 
 ## [3.3.2] - 2020-02-26
 
 Bugfix release.
 
-- Fixes #15, crash caused by battery widget when some accessories have batteries
-- Fixes #57, colors with dashes in the name not found.
+### Fixed
+
+- #15, crash caused by battery widget when some accessories have batteries
+- #57, colors with dashes in the name not found.
 - Also, cjbassi/gotop#127 and cjbassi/gotop#130 were released back in v3.1.0.
 
 ## [3.3.1] - 2020-02-18